+++ src/lando/main/models/jobs.py
@@ -257,24 +257,3 @@ class BaseJob(BaseModel):
             job_dict["repository"] = self.target_repo.short_name
 
         return job_dict
-
-
-class JobAdmin(admin.ModelAdmin):
-    """A base admin class for jobs."""
-
-    list_display = (
-        "id",
-        "status",
-        "target_repo__name",
-        "created_at",
-        "requester_email",
-        "duration_seconds",
-    )
-    list_filter = ("target_repo__name", "requester_email", "created_at")
-    readonly_fields = (
-        "attempts",
-        "duration_seconds",
-        "error",
-        "landed_commit_id",
-        "requester_email",
-    )
